requirements.txt
playsound
python-dotenv
requests